requirements.txt
scipy==1.7.3
matplotlib==3.5.3
plotly==5.14.0
numpy==1.21.6
pandas==1.3.5
streamlit==1.23.1
Pillow==10.1.0
scikit_learn==1.3.2
